+++ src/close-anchor-branch/close-anchor-branch.js
@@ -1,0 +1,6 @@
+import elementReady from 'element-ready';
+
+export default async function closeAnchorBranch() {
+    const closeAnchorBranchCheckbox = await elementReady('#id_close_anchor_branch:not(:disabled)');
+    closeAnchorBranchCheckbox.checked = true;
+}